Move personal wallet UI in the `Personal` namespace
