+++ lib/exe/lib/Cardano/Wallet/Application.hs
@@ -167,13 +167,13 @@ import Cardano.Wallet.TokenMetadata
 import Cardano.Wallet.Transaction
     ( TransactionLayer
     )
-import Cardano.Wallet.UI.Html.Pages.Page
+import Cardano.Wallet.UI.Personal.Html.Pages.Page
     ( PageConfig (..)
     )
-import Cardano.Wallet.UI.Html.Pages.Template.Head
+import Cardano.Wallet.UI.Personal.Html.Pages.Template.Head
     ( HeadConfig (..)
     )
-import Cardano.Wallet.UI.Layer
+import Cardano.Wallet.UI.Personal.Layer
     ( UILayer
     , sourceOfNewTip
     )
@@ -245,10 +245,10 @@ import UnliftIO
 import qualified Cardano.Pool.DB.Layer as Pool
 import qualified Cardano.Wallet.Api.Http.Shelley.Server as Server
 import qualified Cardano.Wallet.DB.Layer as Sqlite
-import qualified Cardano.Wallet.UI.API as Ui
+import qualified Cardano.Wallet.UI.Personal.API as Ui
 
-import qualified Cardano.Wallet.UI.Layer as Ui
-import qualified Cardano.Wallet.UI.Server as Ui
+import qualified Cardano.Wallet.UI.Personal.Layer as Ui
+import qualified Cardano.Wallet.UI.Personal.Server as Ui
 import qualified Network.Wai.Handler.Warp as Warp
 import qualified Servant.Server as Servant
