+++ lib/ui/src/Cardano/Wallet/UI/Personal/Html/Htmx.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Cardano.Wallet.UI.Html.Htmx
+module Cardano.Wallet.UI.Personal.Html.Htmx
     ( hxBoost_
     , hxConfirm_
     , hxEncoding_
